requirements.txt
Django~=4.2.13
redis~=5.0.7
celery~=5.4.0
djangorestframework~=3.15.2
APScheduler~=3.10.4